+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change log
 
+## (2017-11-27) ([@thesmartnik][])
+
+- Update `have_broadcasted_to` matcher to support a record as an argument.
+
+See https://github.com/palkan/action-cable-testing/issues/9
+
 ## 0.1.2 (2017-11-14) ([@palkan][])
 
 - Add RSpec shared contexts to switch between adapters.
@@ -15,3 +21,4 @@ See https://github.com/palkan/action-cable-testing/issues/4.
 - Initial version. ([@palkan][])
 
 [@palkan]: https://github.com/palkan
+[@thesmartnik]: https://github.com/thesmartnik